refactor/ rename "nonprofit_orgs" to "organizations"
