+++ src/organizations/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class NonprofitOrgsConfig(AppConfig):
+class OrganizationsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "nonprofit_orgs"
+    name = "organizations"
